+++ TraZer/Tracer/source/tzMaths.cpp
@@ -18,14 +18,14 @@
 #define	IsZero(x)	((x) > -EQN_EPS && (x) < EQN_EPS)
 
 #ifndef CBRT
-#define     cbrt(x)  ((x) > 0.0 ? pow((double)(x), 1.0/3.0) : \
-			  		 ((x) < 0.0 ? -pow((double)-(x), 1.0/3.0) : 0.0))
+#define     cbrt(x)  ((x) > 0.0 ? pow((float)(x), 1.0/3.0) : \
+			  		 ((x) < 0.0 ? -pow((float)-(x), 1.0/3.0) : 0.0))
 #endif
 
 
 int 
-SolveQuadric(double c[3], double s[2]) {
-    double p, q, D;
+SolveQuadric(float c[3], float s[2]) {
+	float p, q, D;
 
     /* normal form: x^2 + px + q = 0 */
 
@@ -39,7 +39,7 @@ SolveQuadric(double c[3], double s[2]) {
 	return 1;
     }
     else if (D > 0) {
-	double sqrt_D = sqrt(D);
+		float sqrt_D = sqrtf(D);
 
 	s[ 0 ] =   sqrt_D - p;
 	s[ 1 ] = - sqrt_D - p;
@@ -51,12 +51,12 @@ SolveQuadric(double c[3], double s[2]) {
 
 
 int 
-SolveCubic(double c[4], double s[3]) {
+SolveCubic(float c[4], float s[3]) {
     int     i, num;
-    double  sub;
-    double  A, B, C;
-    double  sq_A, p, q;
-    double  cb_p, D;
+	float  sub;
+	float  A, B, C;
+	float  sq_A, p, q;
+	float  cb_p, D;
 
     /* normal form: x^3 + Ax^2 + Bx + C = 0 */
 
@@ -68,8 +68,8 @@ SolveCubic(double c[4], double s[3]) {
 	x^3 +px + q = 0 */
 
     sq_A = A * A;
-    p = 1.0/3 * (- 1.0/3 * sq_A + B);
-    q = 1.0/2 * (2.0/27 * A * sq_A - 1.0/3 * A * B + C);
+    p = 1.0f/3.0f * (- 1.0f/3.0f * sq_A + B);
+    q = 1.0f/2.0f * (2.0f/27.0f * A * sq_A - 1.0f/3.0f * A * B + C);
 
     /* use Cardano's formula */
 
@@ -82,25 +82,25 @@ SolveCubic(double c[4], double s[3]) {
 		    num = 1;
 		}
 	else { /* one single and one double solution */
-	    double u = cbrt(-q);
+		float u = cbrtf(-q);
 	    s[ 0 ] = 2 * u;
 	    s[ 1 ] = - u;
 	    num = 2;
 	}
     }
     else if (D < 0) { /* Casus irreducibilis: three real solutions */
-		double phi = 1.0/3 * acos(-q / sqrt(-cb_p));
-		double t = 2 * sqrt(-p);
+		float phi = 1.0f/3.0f * acosf(-q / sqrtf(-cb_p));
+		float t = 2.0f * sqrtf(-p);
 
-		s[ 0 ] =   t * cos(phi);
-		s[ 1 ] = - t * cos(phi + M_PI / 3);
-		s[ 2 ] = - t * cos(phi - M_PI / 3);
+		s[ 0 ] =   t * cosf(phi);
+		s[ 1 ] = - t * cosf(phi + M_PI / 3.0f);
+		s[ 2 ] = - t * cosf(phi - M_PI / 3.0f);
 		num = 3;
     }
     else { /* one real solution */
-		double sqrt_D = sqrt(D);
-		double u = cbrt(sqrt_D - q);
-		double v = - cbrt(sqrt_D + q);
+		float sqrt_D = sqrtf(D);
+		float u = cbrtf(sqrt_D - q);
+		float v = - cbrtf(sqrt_D + q);
 
 		s[ 0 ] = u + v;
 		num = 1;
@@ -108,7 +108,7 @@ SolveCubic(double c[4], double s[3]) {
 
     /* resubstitute */
 
-    sub = 1.0/3 * A;
+    sub = 1.0f/3.0f * A;
 
     for (i = 0; i < num; ++i)
 	s[ i ] -= sub;
@@ -119,11 +119,11 @@ SolveCubic(double c[4], double s[3]) {
 
 
 int 
-SolveQuartic(double c[5], double s[4]) {
-    double  coeffs[4];
-    double  z, u, v, sub;
-    double  A, B, C, D;
-    double  sq_A, p, q, r;
+SolveQuartic(float c[5], float s[4]) {
+	float  coeffs[4];
+	float  z, u, v, sub;
+	float  A, B, C, D;
+	float  sq_A, p, q, r;
     int     i, num;
 
     /* normal form: x^4 + Ax^3 + Bx^2 + Cx + D = 0 */
@@ -137,9 +137,9 @@ SolveQuartic(double c[5], double s[4]) {
 	x^4 + px^2 + qx + r = 0 */
 
     sq_A = A * A;
-    p = - 3.0/8 * sq_A + B;
-    q = 1.0/8 * sq_A * A - 1.0/2 * A * B + C;
-    r = - 3.0/256*sq_A*sq_A + 1.0/16*sq_A*B - 1.0/4*A*C + D;
+    p = - 3.0f/8.0f * sq_A + B;
+    q = 1.0f/8.0f * sq_A * A - 1.0f/2.0f * A * B + C;
+    r = - 3.0f/256.0f*sq_A*sq_A + 1.0f/16.0f*sq_A*B - 1.0f/4.0f*A*C + D;
 
     if (IsZero(r)) {
 		/* no absolute term: y(y^3 + py + q) = 0 */
@@ -156,10 +156,10 @@ SolveQuartic(double c[5], double s[4]) {
     else {
 		/* solve the resolvent cubic ... */
 
-		coeffs[ 0 ] = 1.0/2 * r * p - 1.0/8 * q * q;
+		coeffs[ 0 ] = 1.0f/2.0f * r * p - 1.0f/8.0f * q * q;
 		coeffs[ 1 ] = - r;
-		coeffs[ 2 ] = - 1.0/2 * p;
-		coeffs[ 3 ] = 1;
+		coeffs[ 2 ] = - 1.0f/2.0f * p;
+		coeffs[ 3 ] = 1.0f;
 		
 		(void) SolveCubic(coeffs, s);
 
@@ -175,14 +175,14 @@ SolveQuartic(double c[5], double s[4]) {
 		if (IsZero(u))
 		    u = 0;
 		else if (u > 0)
-		    u = sqrt(u);
+		    u = sqrtf(u);
 		else
 		    return 0;
 
 		if (IsZero(v))
 		    v = 0;
 		else if (v > 0)
-		    v = sqrt(v);
+		    v = sqrtf(v);
 		else
 		    return 0;
 
@@ -201,7 +201,7 @@ SolveQuartic(double c[5], double s[4]) {
 
     /* resubstitute */
 
-    sub = 1.0/4 * A;
+    sub = 1.0f/4.0f * A;
 
     for (i = 0; i < num; ++i)
 		s[ i ] -= sub;
