+++ TraZer/Tracer/source/tzMeshTriangle.cpp
@@ -93,7 +93,7 @@ tzMeshTriangle::get_normal(void) const {
 
 tzBBox
 tzMeshTriangle::get_bounding_box(void) {
-	double delta = 0.0001;  // to avoid degenerate bounding boxes
+	float delta = 0.0001f;  // to avoid degenerate bounding boxes
 	
 	tzPoint3D v1(mesh_ptr->vertices[index0]);
 	tzPoint3D v2(mesh_ptr->vertices[index1]);
@@ -114,24 +114,24 @@ bool tzMeshTriangle::shadowHit(const tzRay& ray, float& tmin) const {
 	tzPoint3D v1(mesh_ptr->vertices[index1]);
 	tzPoint3D v2(mesh_ptr->vertices[index2]);
 
-	double a = v0.x - v1.x, b = v0.x - v2.x, c = ray.d.x, d = v0.x - ray.o.x; 
-	double e = v0.y - v1.y, f = v0.y - v2.y, g = ray.d.y, h = v0.y - ray.o.y;
-	double i = v0.z - v1.z, j = v0.z - v2.z, k = ray.d.z, l = v0.z - ray.o.z;
+	float a = v0.x - v1.x, b = v0.x - v2.x, c = ray.d.x, d = v0.x - ray.o.x;
+	float e = v0.y - v1.y, f = v0.y - v2.y, g = ray.d.y, h = v0.y - ray.o.y;
+	float i = v0.z - v1.z, j = v0.z - v2.z, k = ray.d.z, l = v0.z - ray.o.z;
 		
-	double m = f * k - g * j, n = h * k - g * l, p = f * l - h * j;
-	double q = g * i - e * k, s = e * j - f * i;
+	float m = f * k - g * j, n = h * k - g * l, p = f * l - h * j;
+	float q = g * i - e * k, s = e * j - f * i;
 	
-	double inv_denom  = 1.0 / (a * m + b * q + c * s);
+	float inv_denom  = 1.0f / (a * m + b * q + c * s);
 	
-	double e1 = d * m - b * n - c * p;
-	double beta = e1 * inv_denom;
+	float e1 = d * m - b * n - c * p;
+	float beta = e1 * inv_denom;
 	
-	if (beta < 0.0)
+	if (beta < 0.0f)
 	 	return (false);
 	
-	double r = r = e * l - h * i;
-	double e2 = a * n + d * q + c * r;
-	double gamma = e2 * inv_denom;
+	float r = r = e * l - h * i;
+	float e2 = a * n + d * q + c * r;
+	float gamma = e2 * inv_denom;
 	
 	if (gamma < 0.0 )
 	 	return (false);
@@ -139,8 +139,8 @@ bool tzMeshTriangle::shadowHit(const tzRay& ray, float& tmin) const {
 	if (beta + gamma > 1.0)
 		return (false);
 			
-	double e3 = a * p - b * r + d * s;
-	double t = e3 * inv_denom;
+	float e3 = a * p - b * r + d * s;
+	float t = e3 * inv_denom;
 	
 	if (t < kEpsilon) 
 		return (false);
