+++ TraZer/Tracer/source/tzSmoothTriangle.cpp
@@ -90,11 +90,11 @@ tzSmoothTriangle::interpolate_normal(const float beta, const float gamma) const 
 
 tzBBox
 tzSmoothTriangle::get_bounding_box(void) {
-	double delta = 0.0001; 
+	float delta = 0.0001f;
 	
-	return(tzBBox(fmin(fmin(v0.x, v1.x), v2.x) - delta, fmax(fmax(v0.x, v1.x), v2.x) + delta, 
-				fmin(fmin(v0.y, v1.y), v2.y) - delta, fmax(fmax(v0.y, v1.y), v2.y) + delta, 
-				fmin(fmin(v0.z, v1.z), v2.z) - delta, fmax(fmax(v0.z, v1.z), v2.z) + delta));
+	return(tzBBox(fminf(fminf(v0.x, v1.x), v2.x) - delta, fmaxf(fmaxf(v0.x, v1.x), v2.x) + delta, 
+				fminf(fminf(v0.y, v1.y), v2.y) - delta, fmaxf(fmaxf(v0.y, v1.y), v2.y) + delta, 
+				fminf(fminf(v0.z, v1.z), v2.z) - delta, fmaxf(fmaxf(v0.z, v1.z), v2.z) + delta));
 }
 
 
@@ -102,25 +102,26 @@ tzSmoothTriangle::get_bounding_box(void) {
 // ------------------------------------------------------------------------------ hit
 
 bool 
-tzSmoothTriangle::hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const {	
-	double a = v0.x - v1.x, b = v0.x - v2.x, c = ray.d.x, d = v0.x - ray.o.x; 
-	double e = v0.y - v1.y, f = v0.y - v2.y, g = ray.d.y, h = v0.y - ray.o.y;
-	double i = v0.z - v1.z, j = v0.z - v2.z, k = ray.d.z, l = v0.z - ray.o.z;
+tzSmoothTriangle::hit(const tzRay& ray, float& tmin, tzShadeRec& sr) const 
+{
+	float a = v0.x - v1.x, b = v0.x - v2.x, c = ray.d.x, d = v0.x - ray.o.x;
+	float e = v0.y - v1.y, f = v0.y - v2.y, g = ray.d.y, h = v0.y - ray.o.y;
+	float i = v0.z - v1.z, j = v0.z - v2.z, k = ray.d.z, l = v0.z - ray.o.z;
 		
-	double m = f * k - g * j, n = h * k - g * l, p = f * l - h * j;
-	double q = g * i - e * k, s = e * j - f * i;
+	float m = f * k - g * j, n = h * k - g * l, p = f * l - h * j;
+	float q = g * i - e * k, s = e * j - f * i;
 	
-	double inv_denom  = 1.0 / (a * m + b * q + c * s);
+	float inv_denom  = 1.0f / (a * m + b * q + c * s);
 	
-	double e1 = d * m - b * n - c * p;
-	double beta = e1 * inv_denom;
+	float e1 = d * m - b * n - c * p;
+	float beta = e1 * inv_denom;
 	
 	if (beta < 0.0)
 	 	return (false);
 	
-	double r = r = e * l - h * i;
-	double e2 = a * n + d * q + c * r;
-	double gamma = e2 * inv_denom;
+	float r = r = e * l - h * i;
+	float e2 = a * n + d * q + c * r;
+	float gamma = e2 * inv_denom;
 	
 	if (gamma < 0.0)
 	 	return (false);
@@ -128,8 +129,8 @@ tzSmoothTriangle::hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const {
 	if (beta + gamma > 1.0)
 		return (false);
 			
-	double e3 = a * p - b * r + d * s;
-	double t = e3 * inv_denom;
+	float e3 = a * p - b * r + d * s;
+	float t = e3 * inv_denom;
 	
 	if (t < kEpsilon)
 		return (false);
@@ -146,26 +147,26 @@ tzSmoothTriangle::hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const {
 // ------------------------------------------------------------------------------ shadow_hit
 // Hit function for shadow rays
 
-bool tzSmoothTriangle::shadowHit(const tzRay& ray, double& tmin) const 
+bool tzSmoothTriangle::shadowHit(const tzRay& ray, float& tmin) const
 {
-	double a = v0.x - v1.x, b = v0.x - v2.x, c = ray.d.x, d = v0.x - ray.o.x; 
-	double e = v0.y - v1.y, f = v0.y - v2.y, g = ray.d.y, h = v0.y - ray.o.y;
-	double i = v0.z - v1.z, j = v0.z - v2.z, k = ray.d.z, l = v0.z - ray.o.z;
+	float a = v0.x - v1.x, b = v0.x - v2.x, c = ray.d.x, d = v0.x - ray.o.x;
+	float e = v0.y - v1.y, f = v0.y - v2.y, g = ray.d.y, h = v0.y - ray.o.y;
+	float i = v0.z - v1.z, j = v0.z - v2.z, k = ray.d.z, l = v0.z - ray.o.z;
 		
-	double m = f * k - g * j, n = h * k - g * l, p = f * l - h * j;
-	double q = g * i - e * k, s = e * j - f * i;
+	float m = f * k - g * j, n = h * k - g * l, p = f * l - h * j;
+	float q = g * i - e * k, s = e * j - f * i;
 	
-	double inv_denom  = 1.0 / (a * m + b * q + c * s);
+	float inv_denom  = 1.0f / (a * m + b * q + c * s);
 	
-	double e1 = d * m - b * n - c * p;
-	double beta = e1 * inv_denom;
+	float e1 = d * m - b * n - c * p;
+	float beta = e1 * inv_denom;
 	
 	if (beta < 0.0)
 	 	return (false);
 	 	
-	double r = e * l - h * i;
-	double e2 = a * n + d * q + c * r;
-	double gamma = e2 * inv_denom;
+	float r = e * l - h * i;
+	float e2 = a * n + d * q + c * r;
+	float gamma = e2 * inv_denom;
 	
 	if (gamma < 0.0)
 	 	return (false);
@@ -173,8 +174,8 @@ bool tzSmoothTriangle::shadowHit(const tzRay& ray, double& tmin) const
 	if (beta + gamma > 1.0)
 		return (false);
 			
-	double e3 = a * p - b * r + d * s;
-	double t = e3 * inv_denom;
+	float e3 = a * p - b * r + d * s;
+	float t = e3 * inv_denom;
 	
 	if (t < kEpsilon)
 		return (false);
